Creation of functions height, treeCalculateFatBal of test_example_2
The height function calculates the height of a binary tree. The treeCalculateFatBal function calculates the fat bal of each node and stores it in the fatbal variable of the respective node. Finally, the test_example_2 function is a test on the created functions. View in the telegram group the schematics I made as a reference for example 1 of the teacher's pdf.